trading_bot/requirements.txt
pandas==0.24.0
numpy==1.15.1
google-cloud-bigquery==1.9.0
pyarrow==0.12.0
requests==2.22.0
google-cloud-storage==1.13.2
google-cloud-secret-manager==0.2.0
scipy==1.2.1
joblib==0.14.1
alpaca-trade-api==0.41
scikit-learn==0.22.2.post1
ta==0.5.11